New Tag: "Argued By" which is used to hyperref in into OGSN output
In safety projects, were traceability between the safety requirements and the
safety arguementation is necessary it shall be possible provide this
traceability by hyperreferences. Here the chosen tool for the safety
argumentation is the OGSN (OSADL goal-structuring notation editor). The
Argued By tag shall be used to create hyperreferences into the OGSN output.

Signed-off-by: Andreas Platschek <andreas.platschek@opentech.at>